app/config.py
# app/config.py

import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
GOOGLE_CALENDAR_TOKEN = os.getenv("GOOGLE_CALENDAR_TOKEN")
GOOGLE_GMAIL_TOKEN = os.getenv("GOOGLE_GMAIL_TOKEN")
