requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
openai
pydantic
python-dotenv
tqdm
httpx
typing-extensions
